rename events_utils to envvars
